wip: day 5 part 2

+++ src/day6/Day6Part1.cpp
@@ -38,7 +38,7 @@ int Day6Part1::solve()
 			double root2 = (-b - std::sqrt(determinant)) / 2 * a;
 			double r1 = std::ceil(root1);
 			double r2 = std::floor(root2);
-			waysToVictory.push_back(std::abs(r2 - r1 + 1));
+			waysToVictory.push_back(static_cast<int>(std::abs(r2 - r1 + 1)));
 		}
 		else
 		{
